+++ test/test_stem.rb
@@ -3,30 +3,11 @@
 require_relative 'helper'
 
 
-# TODO JGrouper::Stem::TestCase
-class TestStem < Test::Unit::TestCase
-
-  def setup
-    @json = { 'json_class'   => JGrouper::Stem,
-              'display_name' => '',
-              'name'         => '',
-              'uuid'         => 'some-fake-uuid'
-            }.to_json
-    @root      = OpenStruct.new :getDisplayName => '',
-                                :getName        => '',
-                                :getUuid        => 'some-fake-uuid'
-    @stem      = OpenStruct.new :getDisplayName => 'foo',
-                                :getName        => 'foo',
-                                :getUuid        => 'stem-foo'
-    @grandstem = OpenStruct.new :getDisplayName => 'foo:Bar',
-                                :getName        => 'foo:bar',
-                                :getUuid        => 'stem-foo-bar'
-  end
-
+class JGrouper::Stem::TestCase
 
   def test_mocked_find_root_stem
     Java::EduInternet2MiddlewareGrouper::GrouperSession.expects(:startRootSession).returns true
-    Java::EduInternet2MiddlewareGrouper::StemFinder.expects(:findRootStem).returns @root
+    Java::EduInternet2MiddlewareGrouper::StemFinder.expects(:findRootStem).returns grouper_root_stem
 
     JGrouper::Stem.root do |stem|
       assert_not_nil stem
@@ -39,13 +20,13 @@ class TestStem < Test::Unit::TestCase
 
   def test_mocked_stem_to_json
     Java::EduInternet2MiddlewareGrouper::GrouperSession.expects(:startRootSession).returns true
-    Java::EduInternet2MiddlewareGrouper::StemFinder.expects(:findRootStem).returns @root
+    Java::EduInternet2MiddlewareGrouper::StemFinder.expects(:findRootStem).returns grouper_root_stem
 
-    assert_equal @json, JGrouper::Stem.root.to_json
+    assert_equal root_stem_as_json, JGrouper::Stem.root.to_json
   end
 
   def test_stem_from_json
-    stem = JSON.parse @json
+    stem = JSON.parse root_stem_as_json
     assert_not_nil stem
     assert         stem.kind_of? JGrouper::Stem
     assert_equal   '',               stem.display_name
@@ -55,7 +36,7 @@ class TestStem < Test::Unit::TestCase
 
   def test_mocked_find_by_name
     Java::EduInternet2MiddlewareGrouper::GrouperSession.expects(:startRootSession).returns true
-    Java::EduInternet2MiddlewareGrouper::StemFinder.expects(:findByName).returns @stem
+    Java::EduInternet2MiddlewareGrouper::StemFinder.expects(:findByName).returns grouper_stem('foo:bar')
 
     JGrouper::Stem.find('foo:bar') do |stem|
       assert_not_nil stem
@@ -74,8 +55,10 @@ class TestStem < Test::Unit::TestCase
   end
 
   def test_mocked_create_stem_without_display_name_and_without_intermediaries
-    JGrouper::Stem.expects(:root).returns @root
-    @root.expects(:add_stem).returns JGrouper::Stem.send(:from_grouper, @stem)
+    root_stem = grouper_root_stem
+
+    JGrouper::Stem.expects(:root).returns root_stem
+    root_stem.expects(:add_stem).returns JGrouper::Stem.send( :from_grouper, grouper_stem(:foo) )
 
     JGrouper::Stem.create('foo') do |stem|
       assert_not_nil  stem
@@ -87,9 +70,12 @@ class TestStem < Test::Unit::TestCase
   end
 
   def test_mocked_create_stem_with_display_name_and_with_intermediaries
-    JGrouper::Stem.expects(:root).returns @root
-    @root.expects(:add_stem).returns @stem
-    @stem.expects(:add_stem).returns JGrouper::Stem.send(:from_grouper, @grandstem)
+    root_stem = grouper_root_stem
+    stem_foo  = grouper_stem :foo
+
+    JGrouper::Stem.expects(:root).returns root_stem
+    root_stem.expects(:add_stem).returns stem
+    stem.expects(:add_stem).returns JGrouper::Stem.send( :from_grouper, grouper_stem('foo:bar') )
 
     JGrouper::Stem.create('foo:bar', 'foo:Bar') do |stem|
       assert_not_nil  stem